reorganization of functions
